Introduce distgit.repo
Some rpms need to get built for rhel7 and rhel8. Currently, there ius a
special post processsing job that takes the content of the distgits of
the affected rpms, puts it in another target branch, and builds it. This
works, but introduces a new step in building, and the rpm definition is
not as explicit as it could be.

This change allows for a new key in rpm distgit: `repo`. This allows to
override the default distgit name, which comes from the file name.

As the distgit name can theoretically now be overloaded, the key for
locking dependent builds has moved from the qualified (distgit) name to
the qualified key, which is guaranteed to be unique.